parallel support
marks all function parallel safe
add support function for partial aggregate

With postgtes 9.6 parallel queries where introduced.
To let a function run in a parallel worker it needs to
be marked as PARALLEL SAFE.
Futhermore to let an aggregate run partially in a
parallel worker it must be marked as PARALLEL SAFE as well.
As we use an internal state for isotre aggregates there is
an additional need for a combine/serialize/deserialize
support function of that state.